P2/requirements.txt
pymongo[srv]==4.2.0
python-dotenv==0.21.0
PyYAML==6.0